Enable SRBAC by default
This patch enables SRBAC by default instead of requiring explicit
enabling like we did in OSP17.

With this it will have the same behavior as OSP17 but without the need
of policy files that explicitly define this behavior.

Only the API service will include the configuration since policies are
to be checked at the endpoint in Cinder API.

Jira: OSPRH-3281