Samples: Improve LVM
This patch improves the LVM samples so it is closer to what a real
deployment should use.

To do this the human operator will need to label one of the nodes with
"openstack.org/cinder-lvm", and this patch modifies the samples so
that they

- The cinder volume pod can only run on this labeled node.

- The systemd unit that creates and presents the LVM VG it only really
  runs in the labeled node.

Appropriate documentation has been added under the
"config/samples/backends/lvm/" directory in a "README.md" file.