Fix LVM backend
This patch fixes a couple of issues with the LVM backend:

- When the cinder-volume service is restarted for any reason
  (configuration change, update, etc) all connections to attached
  volumes will be broken and will never be recovered.

- At least with NVMe-oF (haven't checked with iSCSI), if a volume has
  been attached at any point and the cinder-volume service is restarted,
  then all following attachments won't work.

This patch fixes this issue by using the host network when using the LVM
backend.

Only the cinder-volume service is launched using the host's network
namespace so that the probes can still be run in an isolated network
namespaces; otherwise we would either need to disable the probes or
change the probes for LVM backends.