Always init conditions to Unknown
At the beginning of the reconcile we should always set all conditions to
Unknown to make sure that they are reset to the real status.

This is the right thing to do because we are doing the steps
sequentially, so any step we don't reach, we don't know its status,
hence the "Unknown" status condition.