Use common variable for durations and requeue
We have a number of places where we are using 5 seconds durations, 10
seconds duration, and requeue after 10 seconds.

These are all "magic numbers" spread all over the operator code.

This patch creates 2 constants and 1 variable (because golang doesn't
allow it to be defined as a constant) for these so that they can be used
throughout the code having defined the "magic numbers" in a single
location.