Logs: Don't say Deployment when we mean CRs
There are a couple of places where the logs will say "Deployment
successfully reconciled" when they are actually referring to CRs.

This is quite confusing, because Deployment is an OpenShift object which
we are not using.

This patch changes this so we stop saying Deployment and instead refer
to the specific CR we are talking about.