Only mark ready when reconciliation completes
Fix the global "Ready" condition for Cinder CRs so that it's only set to
"True" when we have gone through all the steps and not just based on the
current value.

The patch also set the global "Ready" condition to False instead of to
Unknown, as that's the case unless proven otherwise by reaching the end
of the reconciliation loop.

This is important for updating deployments, where we may already have a
running pod and the check that that pod is running doesn't mean
that the latest changes have been applied.