Remote useless reconciliation methods
There are a number of placeholder reconciliation method that do nothing
except logging messages of how the method has been entered and exited.

This is useless code as well as clutter in the logs, so this patch
removes them.

Affected methods are:

- reconcileInit
- reconcileUpdate
- reconcileUpgrade

There are instances where reconcileInit is actually doing things, and in
those cases methods were left untouched.