Fix ObservedGeneration assigment
ObservedGeneration needs to be assigned as soon as the reconciler starts
processing the CRD and its conditions have been reset correctly, because
at that point the controller has "observed" the changes and set the
conditions accordingly.

This patch moves the assignment from the end of the reconcile loop to
almost the beginning.