Fix status conditions
This patch leverages the ObservedGeneration to fix the status
conditions.

Top level cinder CRD cannot use the information of the sub CRDs content
unless the Generation and ObservedGeneration match, and it will only be
ready if it's generation is right and all the subCRDs are ready.

For the sub CRDs to be ready they need to:

- Have observed and reacted to the latest Generation
- Have the right number of replicas
- Be deployment ready or have requested 0 replicas and not be deployment
  ready