README cleanup
Our current README.md file does not provide a quick guide to start
paying with OpenStack and a Cinder backend.

In this patch we update the README.md to present a quick starting guide
that deploys OpenStack with LVM as a block device backend.

It also moves a number of sections to their own independent files as a
cleanup of the top level README.