Fix backend label when using non cinder name
When the Cinder CR is created using a name that is not "cinder" the
`backend` label in the StatefulSet and the final cinder volume Pod will
have the wrong value.

This is because the code that determines the backend name in the cinder
volume controller doesn't have access to the actual name, so it tries to
figure it out from the name of the CinderVolume CR, but since the name
no longer starts with `cinder-volume`, then the value is wrong.

This patch changes the way we do this, and now the cinder-operator adds
a label to the CinderVolume CR with the name of the backend, then the
cinder volume controller uses it to pass it to the statefulset .

Jira: OSPRH-7396