[validation] check cinderVolume names are valid
CinderVolume name is <cinder name>-volume-<volume name>
The CinderVolume controller creates StatefulSet for volume service to run.
This adds a StatefulSet pod's label
"controller-revision-hash": "<statefulset_name>-<hash>"
to the pod.
The kubernetes label is restricted under 63 char and the revision
hash is an int32, 10 chars + the hyphen. Which results in a default
statefulset max len of 52 chars. The statefulset name also
contain the cinder name and -volume-. So the max len also need to
be reduced bye the length of those.

Also the name of the created cinderVolume name match a lowercase
RFC 1123.

Depends-On: https://github.com/openstack-k8s-operators/lib-common/pull/532

Jira: https://issues.redhat.com/browse/OSPRH-8063

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>