Move to VerifySecret when checking the ctlplane secret
Instead of recording a hash of the entire ctlplane secret, a new
function is introduced that uses lib-common's VerifySecret() to
verify and return the hash of only the cinder service's portion
of the secret. This avoids restarting all cinder services when
an unrelated portion of the ctlplane secret changes.

This change follows a similar manila PR [1].

[1] https://github.com/openstack-k8s-operators/manila-operator/pull/324

Jira: OSPRH-8192