Enable notifications for Cinder
This patch introduces a new "notificationInstanceBus" parameter in the
Cinder CR. This parameter is used to create a TransportURL that will
be used for notification purposes. When a notificationInstanceBus
parameter is set, cinder controller:
1. initializes the Notification condition provided by lib-common
2. requests a new TransportURL, in case of a dedicated rabbitmq
   instance, and it follows the known transportURL pattern
3. fetches from the resulting k8s Secret the info required to customize
   00-config.conf, where oslo_messaging_notifications is present

This patch also adds test coverage for notifications. In particular it
verifies that the .Status CR field is properly updated, and Cinder is
connected to the same rabbitmq instance. It then checks that the spec
is properly updated when notifications are disabled (by removing the
notificationBusInstance parameter from the Spec) or the parameter
switches to a dedicated rabbitmq instance.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>