Update oslo_reports file_event_handler
With PR [1] root has been removed from cinder. As a consequence
is not possible to trigger gmr with a touch on /etc/cinder path
anymore [2] because it belongs to the root user.
This patch updates the file_event_handler path to point to /var/lib/cinder
that we know is chowned to cinder:cinder by [3].

[1] https://github.com/openstack-k8s-operators/cinder-operator/pull/417
[2] https://docs.openstack.org/cinder/latest/contributor/gmr.html
[3] https://github.com/openstack-k8s-operators/tcib/blob/main/container-images/kolla/base/uid_gid_manage.sh#L40

Jira: https://issues.redhat.com/browse/OSPRH-17882

Signed-off-by: Francesco Pantano <fpantano@redhat.com>